[feedback] refactor to FeedbackActuatorBase / cover subpackage
Mirrors the time_based and endstop refactors (PRs 2/5 and 3/5):
extracts the feedback-cover control logic into a reusable
FeedbackActuatorBase that drives any actuator entity through the
IActuator interface, and moves the cover platform into a subpackage.

New layout:
  feedback/__init__.py            — FeedbackActuatorBase, FEEDBACK_ACTUATOR_SCHEMA,
                                    apply_feedback_actuator_config helpers
  feedback/feedback_actuator.h    — shared base class (Component + IActuator*)
  feedback/feedback_actuator.cpp  — shared loop / control / sensor handling
  feedback/cover/__init__.py      — thin cover platform, AUTO_LOAD parent
  feedback/cover/feedback_cover.h — FeedbackCover : FeedbackActuatorBase, Cover
  feedback/cover/feedback_cover.cpp — get_traits + dump_config

No user-visible behaviour change. Existing YAML configurations
continue to work unchanged.
