+++ esphome/components/feedback/feedback_actuator.cpp
@@ -1,0 +1,409 @@
+#include "feedback_actuator.h"
+#include "esphome/core/hal.h"
+#include "esphome/core/log.h"
+#include "esphome/core/application.h"
+
+namespace esphome::feedback {
+
+static const char *const TAG = "feedback";
+
+static constexpr uint32_t DIRECTION_CHANGE_TIMEOUT_ID = 1;
+
+using namespace esphome::actuator;
+
+void FeedbackActuatorBase::setup() {
+  auto restored = this->actuator_->do_restore_state();
+  if (!restored.has_value()) {
+    // if no other information, assume half open
+    this->actuator_->set_position(0.5f);
+  }
+  this->actuator_->set_operation(ACTUATOR_OPERATION_IDLE);
+
+#ifdef USE_BINARY_SENSOR
+  // if available, get position from endstop sensors
+  if (this->open_endstop_ != nullptr && this->open_endstop_->state) {
+    this->actuator_->set_position(ACTUATOR_OPEN);
+  } else if (this->close_endstop_ != nullptr && this->close_endstop_->state) {
+    this->actuator_->set_position(ACTUATOR_CLOSED);
+  }
+
+  // if available, get moving state from sensors
+  if (this->open_feedback_ != nullptr && this->open_feedback_->state) {
+    this->actuator_->set_operation(ACTUATOR_OPERATION_OPENING);
+  } else if (this->close_feedback_ != nullptr && this->close_feedback_->state) {
+    this->actuator_->set_operation(ACTUATOR_OPERATION_CLOSING);
+  }
+#endif
+
+  this->last_recompute_time_ = this->start_dir_time_ = App.get_loop_component_start_time();
+}
+
+#ifdef USE_BINARY_SENSOR
+
+void FeedbackActuatorBase::set_open_sensor(binary_sensor::BinarySensor *open_feedback) {
+  this->open_feedback_ = open_feedback;
+
+  // setup callbacks to react to sensor changes
+  open_feedback->add_on_state_callback([this](bool state) {
+    ESP_LOGD(TAG, "'%s' - Open feedback '%s'.", this->actuator_->get_entity_name(), state ? "STARTED" : "ENDED");
+    this->recompute_position_();
+    if (!state && this->infer_endstop_ && this->current_trigger_operation_ == ACTUATOR_OPERATION_OPENING) {
+      this->endstop_reached_(true);
+    }
+    this->set_current_operation_(state ? ACTUATOR_OPERATION_OPENING : ACTUATOR_OPERATION_IDLE, false);
+  });
+}
+
+void FeedbackActuatorBase::set_close_sensor(binary_sensor::BinarySensor *close_feedback) {
+  this->close_feedback_ = close_feedback;
+
+  close_feedback->add_on_state_callback([this](bool state) {
+    ESP_LOGD(TAG, "'%s' - Close feedback '%s'.", this->actuator_->get_entity_name(), state ? "STARTED" : "ENDED");
+    this->recompute_position_();
+    if (!state && this->infer_endstop_ && this->current_trigger_operation_ == ACTUATOR_OPERATION_CLOSING) {
+      this->endstop_reached_(false);
+    }
+
+    this->set_current_operation_(state ? ACTUATOR_OPERATION_CLOSING : ACTUATOR_OPERATION_IDLE, false);
+  });
+}
+
+void FeedbackActuatorBase::set_open_endstop(binary_sensor::BinarySensor *open_endstop) {
+  this->open_endstop_ = open_endstop;
+  open_endstop->add_on_state_callback([this](bool state) {
+    if (state) {
+      this->endstop_reached_(true);
+    }
+  });
+}
+
+void FeedbackActuatorBase::set_close_endstop(binary_sensor::BinarySensor *close_endstop) {
+  this->close_endstop_ = close_endstop;
+  close_endstop->add_on_state_callback([this](bool state) {
+    if (state) {
+      this->endstop_reached_(false);
+    }
+  });
+}
+
+void FeedbackActuatorBase::set_close_obstacle_sensor(binary_sensor::BinarySensor *close_obstacle) {
+  this->close_obstacle_ = close_obstacle;
+
+  close_obstacle->add_on_state_callback([this](bool state) {
+    if (state && (this->actuator_->get_operation() == ACTUATOR_OPERATION_CLOSING ||
+                  this->current_trigger_operation_ == ACTUATOR_OPERATION_CLOSING)) {
+      ESP_LOGD(TAG, "'%s' - Close obstacle detected.", this->actuator_->get_entity_name());
+      this->start_direction_(ACTUATOR_OPERATION_IDLE);
+
+      if (this->obstacle_rollback_) {
+        this->target_position_ =
+            clamp(this->actuator_->get_position() + this->obstacle_rollback_, ACTUATOR_CLOSED, ACTUATOR_OPEN);
+        this->start_direction_(ACTUATOR_OPERATION_OPENING);
+      }
+    }
+  });
+}
+
+void FeedbackActuatorBase::set_open_obstacle_sensor(binary_sensor::BinarySensor *open_obstacle) {
+  this->open_obstacle_ = open_obstacle;
+
+  open_obstacle->add_on_state_callback([this](bool state) {
+    if (state && (this->actuator_->get_operation() == ACTUATOR_OPERATION_OPENING ||
+                  this->current_trigger_operation_ == ACTUATOR_OPERATION_OPENING)) {
+      ESP_LOGD(TAG, "'%s' - Open obstacle detected.", this->actuator_->get_entity_name());
+      this->start_direction_(ACTUATOR_OPERATION_IDLE);
+
+      if (this->obstacle_rollback_) {
+        this->target_position_ =
+            clamp(this->actuator_->get_position() - this->obstacle_rollback_, ACTUATOR_CLOSED, ACTUATOR_OPEN);
+        this->start_direction_(ACTUATOR_OPERATION_CLOSING);
+      }
+    }
+  });
+}
+#endif
+
+void FeedbackActuatorBase::endstop_reached_(bool open_endstop) {
+  const uint32_t now = App.get_loop_component_start_time();
+
+  this->actuator_->set_position(open_endstop ? ACTUATOR_OPEN : ACTUATOR_CLOSED);
+
+  // only act if endstop activated while moving in the right direction, in case we are coming back
+  // from a position slightly past the endpoint
+  if (this->current_trigger_operation_ == (open_endstop ? ACTUATOR_OPERATION_OPENING : ACTUATOR_OPERATION_CLOSING)) {
+    float dur = (now - this->start_dir_time_) / 1e3f;
+    ESP_LOGD(TAG, "'%s' - %s endstop reached. Took %.1fs.", this->actuator_->get_entity_name(),
+             open_endstop ? "Open" : "Close", dur);
+
+    // if there is no external mechanism, stop the actuator
+    if (!this->has_built_in_endstop_) {
+      this->start_direction_(ACTUATOR_OPERATION_IDLE);
+    } else {
+      this->set_current_operation_(ACTUATOR_OPERATION_IDLE, true);
+    }
+  }
+
+  // always sync position and publish
+  this->actuator_->do_publish_state(true);
+  this->last_publish_time_ = now;
+}
+
+void FeedbackActuatorBase::set_current_operation_(ActuatorOperation operation, bool is_triggered) {
+  if (is_triggered) {
+    this->current_trigger_operation_ = operation;
+  }
+
+  // if it is setting the actual operation (not triggered one) or
+  // if we don't have moving sensor, we operate in optimistic mode, assuming actions take place immediately
+  // thus, triggered operation always sets current operation.
+  // otherwise, current operation comes from sensor, and may differ from requested operation
+  // this might be from delays or complex actions, or because the movement was not trigger by the component
+  // but initiated externally
+
+#ifdef USE_BINARY_SENSOR
+  if (!is_triggered || (this->open_feedback_ == nullptr || this->close_feedback_ == nullptr))
+#endif
+  {
+    const uint32_t now = App.get_loop_component_start_time();
+    this->actuator_->set_operation(operation);
+    this->start_dir_time_ = this->last_recompute_time_ = now;
+    this->actuator_->do_publish_state(true);
+    this->last_publish_time_ = now;
+  }
+}
+
+void FeedbackActuatorBase::loop() {
+  if (this->actuator_->get_operation() == ACTUATOR_OPERATION_IDLE)
+    return;
+  const uint32_t now = App.get_loop_component_start_time();
+
+  // Recompute position every loop cycle
+  this->recompute_position_();
+
+  // if we initiated the move, check if we reached position or max time
+  // (stoping from endstop sensor is handled in callback)
+  if (this->current_trigger_operation_ != ACTUATOR_OPERATION_IDLE) {
+    if (this->is_at_target_()) {
+      if (this->has_built_in_endstop_ &&
+          (this->target_position_ == ACTUATOR_OPEN || this->target_position_ == ACTUATOR_CLOSED)) {
+        // Don't trigger stop, let the actuator stop by itself.
+        this->set_current_operation_(ACTUATOR_OPERATION_IDLE, true);
+      } else {
+        this->start_direction_(ACTUATOR_OPERATION_IDLE);
+      }
+    } else if (now - this->start_dir_time_ > this->max_duration_) {
+      ESP_LOGD(TAG, "'%s' - Max duration reached. Stopping.", this->actuator_->get_entity_name());
+      this->start_direction_(ACTUATOR_OPERATION_IDLE);
+    }
+  }
+
+  // update current position at requested interval, regardless of who started the movement
+  // so that we also update UI if there was an external movement
+  // don't save intermediate positions
+  if (now - this->last_publish_time_ > this->update_interval_) {
+    this->actuator_->do_publish_state(false);
+    this->last_publish_time_ = now;
+  }
+}
+
+void FeedbackActuatorBase::control_(const ActuatorCallBase &call) {
+  // stop action logic
+  if (call.get_stop()) {
+    this->start_direction_(ACTUATOR_OPERATION_IDLE);
+  } else if (call.get_toggle().has_value()) {
+    // toggle action logic: OPEN - STOP - CLOSE
+    if (this->current_trigger_operation_ != ACTUATOR_OPERATION_IDLE) {
+      this->start_direction_(ACTUATOR_OPERATION_IDLE);
+    } else {
+      if (this->actuator_->get_position() == ACTUATOR_CLOSED || this->last_operation_ == ACTUATOR_OPERATION_CLOSING) {
+        this->target_position_ = ACTUATOR_OPEN;
+        this->start_direction_(ACTUATOR_OPERATION_OPENING);
+      } else {
+        this->target_position_ = ACTUATOR_CLOSED;
+        this->start_direction_(ACTUATOR_OPERATION_CLOSING);
+      }
+    }
+  } else {
+    auto pos_opt = call.get_position();
+    if (!pos_opt.has_value())
+      return;
+    // go to position action
+    auto pos = *pos_opt;
+    if (pos == this->actuator_->get_position()) {
+      // already at target,
+
+      // for actuators with built in end stop, if we don't have sensors we should send the command again
+      // to make sure the assumed state is not wrong
+      if (this->has_built_in_endstop_ && ((pos == ACTUATOR_OPEN
+#ifdef USE_BINARY_SENSOR
+                                           && this->open_endstop_ == nullptr
+#endif
+                                           && !this->infer_endstop_) ||
+                                          (pos == ACTUATOR_CLOSED
+#ifdef USE_BINARY_SENSOR
+                                           && this->close_endstop_ == nullptr
+#endif
+                                           && !this->infer_endstop_))) {
+        this->target_position_ = pos;
+        this->start_direction_(pos == ACTUATOR_CLOSED ? ACTUATOR_OPERATION_CLOSING : ACTUATOR_OPERATION_OPENING);
+      } else if (this->actuator_->get_operation() != ACTUATOR_OPERATION_IDLE ||
+                 this->current_trigger_operation_ != ACTUATOR_OPERATION_IDLE) {
+        // if we are moving, stop
+        this->start_direction_(ACTUATOR_OPERATION_IDLE);
+      }
+    } else {
+      this->target_position_ = pos;
+      this->start_direction_(pos < this->actuator_->get_position() ? ACTUATOR_OPERATION_CLOSING
+                                                                   : ACTUATOR_OPERATION_OPENING);
+    }
+  }
+}
+
+void FeedbackActuatorBase::stop_prev_trigger_() {
+  if (this->direction_change_waittime_.has_value()) {
+    this->cancel_timeout(DIRECTION_CHANGE_TIMEOUT_ID);
+  }
+  if (this->prev_command_trigger_ != nullptr) {
+    this->prev_command_trigger_->stop_action();
+    this->prev_command_trigger_ = nullptr;
+  }
+}
+
+bool FeedbackActuatorBase::is_at_target_() const {
+  // if initiated externally, current operation might be different from
+  // operation that was triggered, thus evaluate position against what was asked
+
+  switch (this->current_trigger_operation_) {
+    case ACTUATOR_OPERATION_OPENING:
+      return this->actuator_->get_position() >= this->target_position_;
+    case ACTUATOR_OPERATION_CLOSING:
+      return this->actuator_->get_position() <= this->target_position_;
+    case ACTUATOR_OPERATION_IDLE:
+      return this->actuator_->get_operation() == ACTUATOR_OPERATION_IDLE;
+    default:
+      return true;
+  }
+}
+
+void FeedbackActuatorBase::start_direction_(ActuatorOperation dir) {
+  Trigger<> *trig;
+
+#ifdef USE_BINARY_SENSOR
+  binary_sensor::BinarySensor *obstacle{nullptr};
+#endif
+
+  switch (dir) {
+    case ACTUATOR_OPERATION_IDLE:
+      trig = &this->stop_trigger_;
+      break;
+    case ACTUATOR_OPERATION_OPENING:
+      this->last_operation_ = dir;
+      trig = &this->open_trigger_;
+#ifdef USE_BINARY_SENSOR
+      obstacle = this->open_obstacle_;
+#endif
+      break;
+    case ACTUATOR_OPERATION_CLOSING:
+      this->last_operation_ = dir;
+      trig = &this->close_trigger_;
+#ifdef USE_BINARY_SENSOR
+      obstacle = this->close_obstacle_;
+#endif
+      break;
+    default:
+      return;
+  }
+
+  this->stop_prev_trigger_();
+
+#ifdef USE_BINARY_SENSOR
+  // check if there is an obstacle to start the new operation -> abort without any change
+  // the case when an obstacle appears while moving is handled in the callback
+  if (obstacle != nullptr && obstacle->state) {
+    ESP_LOGD(TAG, "'%s' - %s obstacle detected. Action not started.", this->actuator_->get_entity_name(),
+             dir == ACTUATOR_OPERATION_OPENING ? "Open" : "Close");
+    return;
+  }
+#endif
+
+  // if we are moving and need to move in the opposite direction
+  // check if we have a wait time
+  if (this->direction_change_waittime_.has_value() && dir != ACTUATOR_OPERATION_IDLE &&
+      this->actuator_->get_operation() != ACTUATOR_OPERATION_IDLE && dir != this->actuator_->get_operation()) {
+    const uint32_t waittime = *this->direction_change_waittime_;
+    ESP_LOGD(TAG, "'%s' - Reversing direction.", this->actuator_->get_entity_name());
+    this->start_direction_(ACTUATOR_OPERATION_IDLE);
+    this->set_timeout(DIRECTION_CHANGE_TIMEOUT_ID, waittime, [this, dir]() { this->start_direction_(dir); });
+  } else {
+    this->set_current_operation_(dir, true);
+    this->prev_command_trigger_ = trig;
+    ESP_LOGD(TAG, "'%s' - Firing '%s' trigger.", this->actuator_->get_entity_name(),
+             dir == ACTUATOR_OPERATION_OPENING   ? "OPEN"
+             : dir == ACTUATOR_OPERATION_CLOSING ? "CLOSE"
+                                                 : "STOP");
+    trig->trigger();
+  }
+}
+
+void FeedbackActuatorBase::recompute_position_() {
+  if (this->actuator_->get_operation() == ACTUATOR_OPERATION_IDLE)
+    return;
+
+  const uint32_t now = App.get_loop_component_start_time();
+  float dir;
+  float action_dur;
+  float min_pos;
+  float max_pos;
+
+  // endstop sensors update position from their callbacks, and sets the fully open/close value
+  // If we have endstop, estimation never reaches the fully open/closed state.
+  // but if movement continues past corresponding endstop (inertia), keep the fully open/close state
+
+  switch (this->actuator_->get_operation()) {
+    case ACTUATOR_OPERATION_OPENING:
+      dir = 1.0f;
+      action_dur = this->open_duration_;
+      min_pos = ACTUATOR_CLOSED;
+      max_pos = (
+#ifdef USE_BINARY_SENSOR
+                    this->open_endstop_ != nullptr ||
+#endif
+                    this->infer_endstop_) &&
+                        this->actuator_->get_position() < ACTUATOR_OPEN
+                    ? 0.99f
+                    : ACTUATOR_OPEN;
+      break;
+    case ACTUATOR_OPERATION_CLOSING:
+      dir = -1.0f;
+      action_dur = this->close_duration_;
+      min_pos = (
+#ifdef USE_BINARY_SENSOR
+                    this->close_endstop_ != nullptr ||
+#endif
+                    this->infer_endstop_) &&
+                        this->actuator_->get_position() > ACTUATOR_CLOSED
+                    ? 0.01f
+                    : ACTUATOR_CLOSED;
+      max_pos = ACTUATOR_OPEN;
+      break;
+    default:
+      return;
+  }
+
+  // check if we have an acceleration_wait_time, and remove from position computation
+  if (now - this->start_dir_time_ > this->acceleration_wait_time_) {
+    uint32_t accel_end_time = this->start_dir_time_ + this->acceleration_wait_time_;
+    uint32_t effective_start;
+    if (static_cast<int32_t>(accel_end_time - this->last_recompute_time_) >= 0) {
+      effective_start = accel_end_time;
+    } else {
+      effective_start = this->last_recompute_time_;
+    }
+    float new_pos =
+        this->actuator_->get_position() + dir * (now - effective_start) / (action_dur - this->acceleration_wait_time_);
+    this->actuator_->set_position(clamp(new_pos, min_pos, max_pos));
+  }
+  this->last_recompute_time_ = now;
+}
+
+}  // namespace esphome::feedback
